refactored out the memory module
